% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_knee_compartment_forces.R
\name{analyse_knee_compartment_forces}
\alias{analyse_knee_compartment_forces}
\title{Analyze Knee Compartment Forces}
\usage{
analyse_knee_compartment_forces(
  jr_data,
  fx_col = "walker_knee_r_on_tibia_r_in_tibia_r_fx",
  fy_col = "walker_knee_r_on_tibia_r_in_tibia_r_fy",
  fz_col = "walker_knee_r_on_tibia_r_in_tibia_r_fz",
  mx_col = "walker_knee_r_on_tibia_r_in_tibia_r_mx",
  my_col = "walker_knee_r_on_tibia_r_in_tibia_r_my",
  mz_col = "walker_knee_r_on_tibia_r_in_tibia_r_mz",
  r_med = c(0, 0, -0.02),
  r_lat = c(0, 0, 0.02),
  fmed_name = "fmed_r",
  flat_name = "flat_r",
  append = TRUE
)
}
\arguments{
\item{jr_data}{A data frame containing joint reaction forces and moments.}

\item{fx_col}{Name of the column containing the x-component of the force.}

\item{fy_col}{Name of the column containing the y-component of the force.}

\item{fz_col}{Name of the column containing the z-component of the force.}

\item{mx_col}{Name of the column containing the x-component of the moment.}

\item{my_col}{Name of the column containing the y-component of the moment.}

\item{mz_col}{Name of the column containing the z-component of the moment.}

\item{r_med}{A numeric vector of length 3 representing the medial vector (default is `c(0,0,-0.02)`).}

\item{r_lat}{A numeric vector of length 3 representing the lateral vector (default is `c(0,0, 0.02)`).}

\item{fmed_name}{The prefix for the columns naming the medial forces (default is `"fmed_r"`).}

\item{flat_name}{The prefix for the columns naming the lateral forces (default is `"flat_r"`).}

\item{append}{Logical value indicating whether to append the calculated forces to the input data frame (`TRUE` by default).}
}
\value{
A data frame containing the original input data along with the calculated medial and lateral forces (if `append = TRUE`), or just the calculated forces (if `append = FALSE`).
}
\description{
This function calculates the medial and lateral forces at the knee based on input force and moment data.
It computes the forces using the provided force and moment data and applies the Moore-Penrose pseudo-inverse to solve for the forces in the medial and lateral compartments.
}
