% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cutoff_correction.R
\name{compute_cutoff_correction}
\alias{compute_cutoff_correction}
\title{Compute Corrected Cutoff Frequency for Filter}
\usage{
compute_cutoff_correction(
  filter_passes = 2,
  cutoff_frequency = 20,
  sampling_frequency = 200,
  type = "lowpass"
)
}
\arguments{
\item{filter_passes}{Integer. The number of passes of the filter. Default is 2.}

\item{cutoff_frequency}{Numeric. The cutoff frequency in Hz. Default is 6 for lowpass.}

\item{sampling_frequency}{Numeric. The sampling frequency in Hz. Default is 200.}

\item{type}{Character. The type of filter. Must be either "lowpass" or "highpass". Default is "lowpass".}
}
\value{
Numeric. The corrected cutoff frequency in Hz.
}
\description{
This function computes the corrected cutoff frequency for a lowpass or highpass filter
using the correction factor by David A. Winter. For highpass filters, Robertson's
high-pass modification is applied.
}
