% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cutoff_correction_highpass.R
\name{compute_cutoff_correction_highpass}
\alias{compute_cutoff_correction_highpass}
\title{Compute Corrected Cutoff Frequency for Highpass Filter}
\usage{
compute_cutoff_correction_highpass(
  filter_passes = 2,
  cutoff_frequency = 20,
  sampling_frequency = 200
)
}
\arguments{
\item{filter_passes}{Integer. The number of passes of the filter. Default is 2.}

\item{cutoff_frequency}{Numeric. The cutoff frequency in Hz. Default is 20.}

\item{sampling_frequency}{Numeric. The sampling frequency in Hz. Default is 200.}
}
\value{
Numeric. The corrected cutoff frequency in Hz.
}
\description{
This function computes the corrected cutoff frequency for a highpass filter
using the correction factor by David A. Winter and Robertson's high-pass modification.
}
