% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cutoff_correction_lowpass.R
\name{compute_cutoff_correction_lowpass}
\alias{compute_cutoff_correction_lowpass}
\title{Compute Corrected Cutoff Frequency for Lowpass Filter}
\usage{
compute_cutoff_correction_lowpass(
  filter_passes = 2,
  cutoff_frequency = 6,
  sampling_frequency = 200
)
}
\arguments{
\item{filter_passes}{Integer. The number of passes of the filter. Default is 2.}

\item{cutoff_frequency}{Numeric. The cutoff frequency in Hz. Default is 6.}

\item{sampling_frequency}{Numeric. The sampling frequency in Hz. Default is 200.}
}
\value{
Numeric. The corrected cutoff frequency in Hz.
}
\description{
This function computes the corrected cutoff frequency for a lowpass filter
using the correction factor by David A. Winter.
}
