% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hjc_bell.R
\name{compute_hjc_bell}
\alias{compute_hjc_bell}
\title{Compute Hip Joint Center using Bell and Brand's Method}
\usage{
compute_hjc_bell(
  data,
  LASIS_name = "L.ASIS",
  RASIS_name = "R.ASIS",
  LPSIS_name = "L.PSIS",
  RPSIS_name = "R.PSIS",
  RHJC_name = "R.HJC",
  LHJC_name = "L.HJC",
  append = TRUE
)
}
\arguments{
\item{data}{A data frame containing marker trajectories with columns corresponding to
the names specified by the `LASIS_name`, `RASIS_name`, `LPSIS_name`, and `RPSIS_name` parameters.}

\item{LASIS_name}{A string specifying the column name or pattern for the left anterior superior
iliac spine markers in the data.}

\item{RASIS_name}{A string specifying the column name or pattern for the right anterior superior
iliac spine markers in the data.}

\item{LPSIS_name}{A string specifying the column name or pattern for the left posterior superior
iliac spine markers in the data.}

\item{RPSIS_name}{A string specifying the column name or pattern for the right posterior superior
iliac spine markers in the data.}

\item{RHJC_name}{A string specifying the prefix for the right hip joint center column names in the output.}

\item{LHJC_name}{A string specifying the prefix for the left hip joint center column names in the output.}

\item{append}{Logical. If TRUE, appends the computed hip joint center columns to the original data frame.
If FALSE, returns a list containing the computed hip joint center data frames.}
}
\value{
A data frame with appended hip joint center columns or a list of two data frames
  containing the right and left hip joint centers.
}
\description{
This function computes the right and left hip joint centers based on Bell and Brand's formula.
}
