% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_jc_mid.R
\name{compute_jc_mid}
\alias{compute_jc_mid}
\title{Compute Euclidean Average of Two 3D Vectors}
\usage{
compute_jc_mid(
  data,
  marker_name_1,
  marker_name_2,
  jc_name = "jc",
  append = TRUE
)
}
\arguments{
\item{data}{A data frame containing columns with 3D vectors.}

\item{marker_name_1}{A string specifying the column name or pattern for the first 3D vector in the data frame.}

\item{marker_name_2}{A string specifying the column name or pattern for the second 3D vector in the data frame.}

\item{jc_name}{A string specifying the prefix for the average vector (joint centre) column names in the output.}

\item{append}{Logical. If TRUE, appends the computed average vector as new columns to the original data frame.}
}
\value{
A data frame with the computed average vector either appended to the original data frame 
  or as a separate data frame depending on the `append` argument.
}
\description{
This function computes the Euclidean average of two 3D vectors for each row in the data frame, commonly used to find the joint centre (e.g., knee, ankle, etc.).
}
