% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_muscle_length_anatomical.R
\name{compute_muscle_length_anatomical}
\alias{compute_muscle_length_anatomical}
\title{Compute Anatomical Muscle Lengths from an OpenSim Model}
\usage{
compute_muscle_length_anatomical(
  model_file = "../TestData/Model_SCALED.osim",
  file_output = "muscle_length_anatomical.sto",
  write_file = FALSE,
  return_object = TRUE
)
}
\arguments{
\item{model_file}{Path to the `.osim` model file (OpenSim 4.x format).}

\item{file_output}{Name of the output file to write muscle lengths (optional).}

\item{write_file}{Logical. If `TRUE`, writes output to a `.sto` file.}

\item{return_object}{Logical. If `TRUE`, returns a data frame of muscle lengths.}
}
\value{
If `return_object = TRUE`, returns a data frame with muscle lengths at time = 0.
        Otherwise, returns invisibly.
}
\description{
Extracts muscle-tendon unit (MTU) lengths from a static OpenSim model at its anatomical (neutral) posture.
}
