% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_muscle_volume_scale_factor.R
\name{compute_muscle_volume_scale_factor}
\alias{compute_muscle_volume_scale_factor}
\title{Calculate Strength Scale Factor Based on Height and Mass}
\usage{
compute_muscle_volume_scale_factor(
  height,
  mass,
  height_ref = 1.68,
  mass_ref = 75.337
)
}
\arguments{
\item{height}{A numeric value representing the height of the individual in meters.}

\item{mass}{A numeric value representing the mass of the individual in kilograms.}

\item{height_ref}{A numeric value for the reference height in meters (default is 1.68).}

\item{mass_ref}{A numeric value for the reference mass in kilograms (default is 75.337).}
}
\value{
A numeric value representing the calculated strength scale factor.
}
\description{
This function computes a muscle volume scale factor based on an individual's height 
and mass relative to reference values from Rajagopal et al. (2016) using regression equations
from Handsfield et al.
}
