% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rotation_between_frames.R
\name{compute_rotation_between_frames}
\alias{compute_rotation_between_frames}
\title{Compute Rotation Between Frames}
\usage{
compute_rotation_between_frames(parent, child)
}
\arguments{
\item{parent}{A 3-dimensional array where each slice represents a frame of rotation matrices.}

\item{child}{A 3-dimensional array of the same dimensions as `parent`, representing another set of rotation matrices.}
}
\value{
A 3-dimensional array of the same dimensions as `parent` and `child`, where each slice represents the rotation matrix between corresponding frames.
}
\description{
This function calculates the rotation matrix between corresponding frames in two 3D arrays.
}
