% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_torque_com.R
\name{compute_torque_com}
\alias{compute_torque_com}
\title{Compute Torque at Center of Mass (COM)}
\usage{
compute_torque_com(com_pos_vec, grf_vec, cop_vec)
}
\arguments{
\item{com_pos_vec}{A numeric vector of length 3 representing the position of the center of mass (COM) in 3D space (x, y, z).}

\item{grf_vec}{A numeric vector of length 3 representing the ground reaction force (GRF) in 3D space (x, y, z).}

\item{cop_vec}{A numeric vector of length 3 representing the position of the center of pressure (COP) in 3D space (x, y, z).}
}
\value{
A numeric vector of length 3 representing the torque (tx, ty, tz) about the COM in 3D space.
}
\description{
This function calculates the torque at the center of mass (COM) given the
position of the COM, the ground reaction force (GRF), and the center of pressure (COP).
The torque is computed as the cross product of the vector from the COM to the COP
and the GRF vector.Note that positive means counterclockwise rotation about a given axis
(when the axis is pointing to you).
}
