% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cross_mat_from_vector.R
\name{create_cross_mat_from_vector}
\alias{create_cross_mat_from_vector}
\title{Convert a Vector to a Cross Product Matrix}
\usage{
create_cross_mat_from_vector(r)
}
\arguments{
\item{r}{A numeric vector of length 3, representing the 3D vector.}
}
\value{
A 3x3 numeric matrix representing the cross product matrix of the input vector.
}
\description{
Given a 3D vector `r`, this function computes the corresponding cross product matrix.
}
