% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rotation_matrix.R
\name{create_rotation_matrix}
\alias{create_rotation_matrix}
\title{Create a 3D Rotation Matrix}
\usage{
create_rotation_matrix(angle_deg, axis)
}
\arguments{
\item{angle_deg}{Numeric. The rotation angle in degrees.}

\item{axis}{Character. The axis of rotation: "x", "y", or "z".}
}
\value{
A 3x3 numeric matrix representing the rotation.
}
\description{
Generates a 3x3 rotation matrix for rotating points in 3D space
around the x, y, or z axis by a specified angle in degrees.
}
\examples{
create_rotation_matrix(90, "z")

}
