% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rotation_matrix_from_vector.r
\name{create_rotation_matrix_from_vector}
\alias{create_rotation_matrix_from_vector}
\title{Convert Rotation Vector to Rotation Matrix}
\usage{
create_rotation_matrix_from_vector(theta3, in_degrees = FALSE)
}
\arguments{
\item{theta3}{A numeric vector of length 3, representing the rotation vector. Its direction
specifies the axis of rotation, and its magnitude specifies the angle (in radians by default).}

\item{in_degrees}{Logical. If \code{TRUE}, the elements of \code{theta3} are interpreted as
degrees and converted to radians internally. Default is \code{FALSE}.}
}
\value{
A 3x3 numeric matrix representing the rotation.
}
\description{
Converts a 3-element rotation vector (axis-angle representation) into a 3x3 rotation matrix,
using Rodrigues' rotation formula. Equivalent to Python's
\code{scipy.spatial.transform.Rotation.from_rotvec}.
}
\examples{
create_rotation_matrix_from_vector(c(0.5, 0.4, 0.3))
create_rotation_matrix_from_vector(c(30, 0, 0), in_degrees = TRUE)

}
\references{
\url{https://en.wikipedia.org/wiki/Axis-angle_representation#Rotation_vector}
\url{https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.transform.Rotation.from_rotvec.html}
}
