% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fukuchi_2018.R
\docType{data}
\name{data_fukuchi_2018}
\alias{data_fukuchi_2018}
\title{Example Dataset: data_fukuchi_2018}
\format{
A data frame with 1375620 rows and 6 variables:
\describe{
  \item{stride}{Numeric stride}
  \item{participant}{Character participant}
  \item{task}{Character task}
  \item{variable}{Character variable}
  \item{value}{Numeric value}
  \item{gait_type}{Character gait_type}
}
}
\source{
https://figshare.com/articles/dataset/A_public_data_set_of_overground_and_treadmill_walking_kinematics_and_kinetics_of_healthy_individuals/5722711
}
\usage{
data_fukuchi_2018
}
\description{
An example dataset of joint angles from Fukuchi et al., 2018, released under https://creativecommons.org/licenses/by/4.0/
}
\references{
Fukuchi CA, Fukuchi RK, Duarte M. 2018. A public dataset of overground and treadmill walking kinematics and kinetics in healthy individuals. PeerJ 6:e4640 https://doi.org/10.7717/peerj.4640
}
\keyword{datasets}
