% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verheul_2024.R
\docType{data}
\name{data_verheul_2024}
\alias{data_verheul_2024}
\title{Example Dataset: data_verheul_2024}
\format{
A data frame with 422685 rows and 7 variables:
\describe{
  \item{value}{Numeric value}
  \item{variable}{Character variable}
  \item{trial}{Character trial}
  \item{participant}{Character participant}
  \item{data_type}{Character data_type}
  \item{leg}{Character leg}
  \item{stance}{Numeric variable}
}
}
\source{
https://github.com/JasperVerheul/horizontal-deceleration-forces
}
\usage{
data_verheul_2024
}
\description{
An example dataset from Verheul et al., 2024, released under https://www.apache.org/licenses/LICENSE-2.0
}
\references{
Verheul J, Harper D, Robinson MA. Forces experienced at different levels of the musculoskeletal system during horizontal decelerations. J Sports Sci. 2024 Dec;42(23):2242-2253. doi: 10.1080/02640414.2024.2428086.
}
\keyword{datasets}
