% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_markerset_info.R
\name{extract_markerset_info}
\alias{extract_markerset_info}
\title{Extract Marker Set Information from OpenSim marker set file}
\usage{
extract_markerset_info(
  marker_file = "markerset.xml",
  output_file = "marker_info.csv",
  write_to_file = FALSE,
  return_obj = TRUE
)
}
\arguments{
\item{marker_file}{A character string specifying the path to the marker set XML file.}

\item{output_file}{A character string specifying the path to the output CSV file.}

\item{write_to_file}{A logical indicating whether to write the result to a CSV file. Default is `FALSE`.}

\item{return_obj}{A logical indicating whether to return the data frame as an object. Default is `TRUE`.}
}
\value{
A data frame containing the marker information.
}
\description{
This function extracts information about the markers in an OpenSim marker set file,
including their name, parent body, fixed status, and location. Optionally, the
results can be written to a CSV file.
}
