% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_lowpass_pad_df.R
\name{filter_lowpass_pad_df}
\alias{filter_lowpass_pad_df}
\title{Filter Markers with Padding to Avoid Artifacts}
\usage{
filter_lowpass_pad_df(
  data,
  time_col = "Time",
  exclude_cols = c(1, 2),
  cut_off = 6,
  sample_rate = 200,
  order = 2,
  pad_time = 0.5
)
}
\arguments{
\item{data}{A data frame containing marker data to be filtered.}

\item{time_col}{A string indicating the name of the column containing the time vector.}

\item{exclude_cols}{A vector of column indices (or names) to exclude from filtering. Usually 1 and 2 to remove Time and Frame.}

\item{cut_off}{Numeric. The cut-off frequency for the low-pass filter in Hz.}

\item{sample_rate}{Numeric. The sample rate of the data in Hz.}

\item{order}{Numeric. The order of the Butterworth filter.}

\item{pad_time}{Numeric. The time length to pad the data at the start and end.}

\item{use_spline_padding}{Logical. Whether to use spline interpolation for padding (TRUE) or replicate the first/last rows (FALSE). Default is FALSE.}
}
\value{
A data frame with the filtered marker data, with padding removed.
}
\description{
This function filters the marker data using a low-pass Butterworth filter with padding at the start and end of the data to avoid artifacts.
}
