% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analog_c3d_osim.R
\name{get_analog_c3d_osim}
\alias{get_analog_c3d_osim}
\title{Get Analog Data from C3D File}
\usage{
get_analog_c3d_osim(
  file_input,
  file_output = "Analog.sto",
  write_file = TRUE,
  return_object = FALSE
)
}
\arguments{
\item{file_input}{A string specifying the path to the input C3D file.}

\item{file_output}{A string specifying the path for the output .sto file (default is "Analog.sto").}

\item{write_file}{A logical value indicating whether to write the output to a file (default is TRUE).}

\item{return_object}{A logical value indicating whether to return the analog data object (default is FALSE).}
}
\value{
If `return_object` is TRUE, returns the analog data object; otherwise, returns NULL.
}
\description{
This function reads a C3D file and extracts the analog data, 
optionally writing it to a .sto file and/or returning the 
opensim data table as an object.
}
