% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osim_run_cmd.R
\name{osim_run_cmd}
\alias{osim_run_cmd}
\title{Run OpenSim Command Line Tool}
\usage{
osim_run_cmd(opensim_path = "C:/\\"OpenSim 4.5\\"/bin", setup_file)
}
\arguments{
\item{opensim_path}{A character string specifying the path to the OpenSim bin directory.}

\item{setup_file}{A character string specifying the path to the setup file (.xml) to be run with OpenSim.}
}
\value{
The result of running the OpenSim command.
}
\description{
This function executes an OpenSim command line tool using `opensim-cmd`.
}
