% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osim_viz_model_cmd.R
\name{osim_viz_model_cmd}
\alias{osim_viz_model_cmd}
\title{Visualize OpenSim Model}
\usage{
osim_viz_model_cmd(
  opensim_path = "C:/\\"OpenSim 4.5\\"/bin",
  model_file_path = "Trunk_pelvis_LL.osim"
)
}
\arguments{
\item{opensim_path}{A character string specifying the path to the OpenSim bin directory.}

\item{model_file_path}{A character string specifying the path to the model file (.osim).}
}
\value{
The result of running the OpenSim command.
}
\description{
This function visualizes an OpenSim model using the `opensim-cmd` tool.
}
