% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osim_viz_motion_cmd.R
\name{osim_viz_motion_cmd}
\alias{osim_viz_motion_cmd}
\title{Visualize Motion Using OpenSim}
\usage{
osim_viz_motion_cmd(
  opensim_path = "C:/\\"OpenSim 4.5\\"/bin",
  motion_file_path = "ik.mot",
  model_file_path = "Trunk_pelvis_LL.osim"
)
}
\arguments{
\item{opensim_path}{A character string specifying the path to the OpenSim bin directory.}

\item{motion_file_path}{A character string specifying the path to the motion file (.mot or .sto).}

\item{model_file_path}{A character string specifying the path to the model file (.osim).}
}
\value{
The result of running the OpenSim command.
}
\description{
This function visualizes a motion file using the OpenSim command line tool `opensim-cmd`.
}
