% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osim_viz_trc_cmd.R
\name{osim_viz_trc_cmd}
\alias{osim_viz_trc_cmd}
\title{Visualize Marker Data in OpenSim}
\usage{
osim_viz_trc_cmd(
  opensim_path = "C:/\\"OpenSim 4.5\\"/bin",
  trc_path = "marker.trc"
)
}
\arguments{
\item{opensim_path}{A character string specifying the path to the OpenSim bin directory.}

\item{trc_path}{A character string specifying the path to the TRC file containing the marker data (.trc).}
}
\value{
The result of running the OpenSim command.
}
\description{
This function visualizes marker data in OpenSim using the `opensim-cmd` tool.
}
