% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trc_deprecated.R
\name{read.trc}
\alias{read.trc}
\title{Read a trc file}
\usage{
read.trc()
}
\arguments{
\item{x}{trc file}
}
\value{
Dataframe containing the marker data extracted from trc file
}
\description{
Read a trc file
}
