% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mot_sto.R
\name{read_mot_sto}
\alias{read_mot_sto}
\title{Read OpenSim MOT or STO File as Data Frame}
\usage{
read_mot_sto(filename)
}
\arguments{
\item{filename}{A character string specifying the path to the file to be read.}
}
\value{
A data frame containing the data from the file, starting from the line that includes the word "time".
}
\description{
This function reads an OpenSim MOT or STO file into a data frame. The file is read after locating the line containing the word "time".
}
\examples{
# Read a sample MOT or STO file
df <- read_mot_sto("example.mot")
}
