% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_coords.R
\name{rotate_coords}
\alias{rotate_coords}
\title{Rotate 3D Coordinates Using a Rotation Matrix}
\usage{
rotate_coords(x, y, z, rotation_matrix)
}
\arguments{
\item{x, }{y, z Numeric vectors of equal length representing 3D coordinates.}

\item{rotation_matrix}{A 3x3 numeric matrix used to rotate the coordinates.}
}
\value{
A data frame with rotated coordinates (`x`, `y`, `z`).
}
\description{
Applies a 3x3 rotation matrix to a set of 3D coordinates.
}
\examples{
rot_mat <- create_rotation_matrix(90, "z")
rotate_coords(1, 0, 0, rot_mat)

}
