% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_osim_table.R
\name{rotate_osim_table}
\alias{rotate_osim_table}
\title{Rotate an Opensim Data Table}
\usage{
rotate_osim_table(table_object, axis = c(1, 0, 0), deg = -90)
}
\arguments{
\item{table_object}{An object representing an OpenSim data table.}

\item{axis}{A numeric vector of length 3 specifying the rotation axis (default is c(1, 0, 0)).}

\item{deg}{A numeric value specifying the rotation angle in degrees (default is -90).}
}
\value{
NULL. The function modifies the input `table_object` in place.
}
\description{
This function rotates a specified OpenSim data table around a given axis
by a specified angle in degrees.
}
