% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_external_loads_xml.R
\name{setup_external_loads_xml}
\alias{setup_external_loads_xml}
\title{Setup External Loads XML}
\usage{
setup_external_loads_xml(
  base_file = "external_loads_base.xml",
  leg = "Left",
  body = "calcn_l",
  datafile = "grf.mot",
  setup_filename = "external_loads.xml"
)
}
\arguments{
\item{base_file}{A character string specifying the path to the base XML file. Default is "external_loads_base.xml".}

\item{leg}{A character string specifying the leg (e.g., "Left" or "Right"). Default is "Left".}

\item{body}{A character string specifying the body part (e.g., "calcn_l"). Default is "calcn_l".}

\item{datafile}{A character string specifying the name of the data file. Default is "grf.mot".}

\item{setup_filename}{A character string specifying the name of the output setup XML file. Default is "external_loads.xml".}
}
\value{
The function does not return a value. It writes the modified XML to the specified setup_filename.
}
\description{
This function sets up an XML file for external loads by modifying
a given base XML file with the specified parameters (leg, body,
datafile) and saves the updated XML file to a specified filename.
}
