% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_joint_reaction_xml.R
\name{setup_joint_reaction_xml}
\alias{setup_joint_reaction_xml}
\title{Set Up Joint Reaction XML for Simulation}
\usage{
setup_joint_reaction_xml(
  base_file = "setup_jr_base.xml",
  model_file = "model_scaled.osim",
  replace_force_set = "false",
  force_set_files = "reserve_actuators.xml",
  results_directory = ".",
  output_precision = "12",
  start_time,
  end_time,
  solve_for_equilibrium_for_auxiliary_states = "false",
  maximum_number_of_integrator_steps = "20000",
  maximum_integrator_step_size = "1",
  minimum_integrator_step_size = "1e-05",
  integrator_error_tolerance = "1e-05",
  step_interval = "1",
  in_degrees = "true",
  forces_file = "Forces.sto",
  joint_names = "all",
  apply_on_bodies = "parent",
  express_in_frame = "child",
  external_loads_file = "external_loads.xml",
  coordinates_file = "ik.mot",
  lowpass_cutoff_frequency_for_coordinates = "12",
  setup_filename = "setup_jr.xml"
)
}
\arguments{
\item{base_file}{Character string. The path to the base XML file to be modified (default is "setup_jr_base.xml").}

\item{model_file}{Character string. The path to the model file to be used (default is "model_scaled.osim").}

\item{replace_force_set}{Character string. Whether to replace the force set ("true" or "false", default is "false").}

\item{force_set_files}{Character string. The path to the force set files to be used (default is "reserve_actuators.xml").}

\item{results_directory}{Character string. The directory where results should be saved (default is current directory, ".").}

\item{output_precision}{Character string. The desired precision for output (default is "12").}

\item{start_time}{Numeric. The start time of the analysis (in seconds).}

\item{end_time}{Numeric. The end time of the analysis (in seconds).}

\item{solve_for_equilibrium_for_auxiliary_states}{Character string. Whether to solve for equilibrium for auxiliary states ("true" or "false", default is "false").}

\item{maximum_number_of_integrator_steps}{Numeric. The maximum number of integrator steps (default is 20000).}

\item{maximum_integrator_step_size}{Numeric. The maximum size of each integrator step (default is 1).}

\item{minimum_integrator_step_size}{Numeric. The minimum size of each integrator step (default is 1e-05).}

\item{integrator_error_tolerance}{Numeric. The error tolerance for the integrator (default is 1e-05).}

\item{step_interval}{Numeric. The step interval for the joint reaction analysis (default is "1").}

\item{in_degrees}{Character string. Whether the results should be expressed in degrees ("true" or "false", default is "true").}

\item{forces_file}{Character string. The path to the forces file (default is "Forces.sto").}

\item{joint_names}{Character string. The names of the joints to analyze (default is "all").}

\item{apply_on_bodies}{Character string. Whether to apply the analysis on the parent or child bodies ("parent" or "child", default is "parent").}

\item{express_in_frame}{Character string. The frame in which to express the reaction forces ("child" or "parent", default is "child").}

\item{external_loads_file}{Character string. The path to the external loads file (default is "external_loads.xml").}

\item{coordinates_file}{Character string. The path to the coordinates file (default is "ik.mot").}

\item{lowpass_cutoff_frequency_for_coordinates}{Numeric. The cutoff frequency for lowpass filtering of coordinates (default is "12").}

\item{setup_filename}{Character string. The name of the XML file to save the modified configuration (default is "setup_jr.xml").}
}
\value{
This function does not return any value. It modifies the XML file in place and saves it to the provided `setup_filename`.
}
\description{
This function modifies an XML configuration file for joint reaction analysis. It updates the XML file with
various parameters such as model file, force set files, analysis times, integrator settings, joint names,
forces file, and other configuration parameters related to joint reactions. The modified XML file is then
saved to the specified setup filename.
}
\details{
This function uses the `XML` package to parse and modify the XML configuration file. Each parameter in the function corresponds
to a specific element in the XML, and its value is updated accordingly. The modified XML file is then saved to the specified location.
}
