% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_normalise_deprecated.R
\name{Time_normalise}
\alias{Time_normalise}
\title{Time normalises continuous data}
\usage{
Time_normalise()
}
\arguments{
\item{DF}{a dataframe containing the continuous data that needs to be time normalised to specified number of points. The first column in the data frame must be the time or sample number column in ascending order}

\item{nodes}{numeric value which indicates how many points you would like your data to be resampled to}
}
\value{
A dataframe of nrow(nodes) which contains the resampled data
}
\description{
Takes a dataframe containing continuous data and resamples it to the specified number of points
}
