% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_3D_frame.R
\name{visualise_3D_frame}
\alias{visualise_3D_frame}
\title{Add a 3D Reference Frame to a Plotly Object}
\usage{
visualise_3D_frame(p, reference_frame, origin, scale_factor = 100)
}
\arguments{
\item{p}{A `plotly` object (typically created with `visualise_3D_marker_position()`).}

\item{reference_frame}{A 3x3 numeric matrix. Each column represents a unit direction vector for an axis (X, Y, Z).}

\item{origin}{A numeric vector of length 3 specifying the origin point of the reference frame.}

\item{scale_factor}{Numeric value indicating the length of the axes to draw (default is 100).}
}
\value{
A modified `plotly` object with reference frame axes added.
}
\description{
Adds visual reference axes (X, Y, Z) to an existing 3D `plotly` object.
Each axis is drawn from the specified origin using direction vectors from a 3x3 reference frame matrix.
}
