% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_3D_marker_position.R
\name{visualise_3D_marker_position}
\alias{visualise_3D_marker_position}
\title{Visualize 3D Marker Position}
\usage{
visualise_3D_marker_position(data, selected_time)
}
\arguments{
\item{data}{A dataframe containing time and marker positions with columns formatted as
`Time`, `markername_X`, `markername_Y`, `markername_Z`.}

\item{selected_time}{Numeric value indicating the specific time point at which to visualize the marker positions.}
}
\value{
A `plotly` object representing the 3D scatter plot of marker positions.
}
\description{
Visualizes the 3D positions of markers at a specific time point using interactive 3D plots.
}
