% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_3D_segment_position.R
\name{visualise_3D_segment_position}
\alias{visualise_3D_segment_position}
\title{Visualize 3D Marker Positions and Segments}
\usage{
visualise_3D_segment_position(
  data,
  selected_frame,
  mapping_data,
  marker_col,
  segment
)
}
\arguments{
\item{data}{A data frame containing time and marker positions with columns formatted as
`Frame`, `Time`, `markername_X`, `markername_Y`, `markername_Z`.}

\item{selected_frame}{Numeric value indicating the frame to visualize the marker positions.}

\item{mapping_data}{A data frame containing mapping information, with columns for marker names
and associated body segments.}

\item{marker_col}{Name of the column in `mapping_data` that contains the marker names.}

\item{segment}{Name of the column in `mapping_data` that contains the body segments.}
}
\value{
A `plotly` object representing the 3D scatter plot of marker positions with segment lines.
}
\description{
This function visualizes the 3D positions of markers at a specific frame,
connecting markers belonging to the same body segment with lines to represent
the segments of the body in a 3D space.
}
