% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_3D_frame_marker.R
\name{visualize_3D_marker_frame_position}
\alias{visualize_3D_marker_frame_position}
\title{Visualize 3D Marker Frame Position}
\usage{
visualize_3D_marker_frame_position(
  data,
  selected_frame,
  reference_frame,
  origin
)
}
\arguments{
\item{data}{A data frame containing marker data, including columns with '_X', '_Y', and '_Z' suffixes for marker positions.}

\item{selected_frame}{A numeric value representing the time frame at which to visualize the markers.}

\item{reference_frame}{A 3x3 matrix representing a reference frame (optional). Each column should represent an axis of the reference frame.}

\item{origin}{A 1x3 numeric vector representing the origin point of the reference frame (optional).}
}
\value{
A Plotly 3D scatter plot visualizing the marker positions at the selected time point, with optional reference frame axes.
}
\description{
This function visualizes the 3D positions of markers at a specified time point
from a dataset. It can also optionally display a reference frame.
}
