% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_3D_marker_position.R
\name{visualize_3D_marker_position}
\alias{visualize_3D_marker_position}
\title{Visualize 3D Marker Position}
\usage{
visualize_3D_marker_position(data, selected_time)
}
\arguments{
\item{data}{A dataframe containing time and marker positions with columns formatted as
`time`, `markername_X`, `markername_Y`, `markername_Z`.}

\item{selected_time}{The specific time point at which to visualize the marker positions.}
}
\value{
A `plotly` object representing the 3D scatter plot of marker positions.
}
\description{
Visualizes the 3D positions of markers at a specific time point using interactive 3D plots.
}
