% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_trc_deprecated.R
\name{write.trc}
\alias{write.trc}
\title{Write a trc file}
\usage{
write.trc()
}
\arguments{
\item{x}{dataframe containing data to write to trc. Should be in the same format produced from read.trc()}

\item{unit}{string contain unit of measurement in either "mm" or "m"}

\item{filename}{string identifying the filename to write to file. Must end in .trc}
}
\value{
trc file of filename
}
\description{
Write a trc file
}
