% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_force_mot.R
\name{write_force_mot}
\alias{write_force_mot}
\title{Write GRF Data Frame to File}
\usage{
write_force_mot(data, filename)
}
\arguments{
\item{data}{A data frame containing GRF data. The first column should represent time.}

\item{filename}{A character string specifying the name of the file to which the data should be written. Include the extension .mot for compatability with OpenSim.}
}
\description{
This function writes a GRF (Ground Reaction Force) data frame to a file with a specific header format suitable for use in OpenSim.
}
\examples{
# Create a sample GRF data frame
df <- data.frame(time = 1:10, force = rnorm(10))
# Write the data frame to a file named "grf_data.mot"
write_force_mot(df, "grf_data.mot")
}
