% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerolag_lowpass_filter.R
\name{zerolag_lowpass_filter}
\alias{zerolag_lowpass_filter}
\title{Filter Markers with Padding to Avoid Artifacts}
\usage{
zerolag_lowpass_filter(
  signal,
  cut_off = 6,
  sample_rate = 120,
  order = 2,
  pad_size = 500
)
}
\arguments{
\item{signal}{A numeric vector containing signal to be filtered.}

\item{cut_off}{Numeric. The cut-off frequency for the low-pass filter in Hz.}

\item{sample_rate}{Numeric. The sample rate of the signal in Hz.}

\item{order}{Numeric. The order of the Butterworth filter.}

\item{pad_size}{Numeric. The number of data points to pad at the start and end of the signal.}
}
\value{
A numeric vector with the filtered signal, with padding removed.
}
\description{
This function filters the signal using a low-pass Butterworth filter with padding at the start and end of the data to avoid artifacts.
}
