% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_joint_angle.R
\name{compute_joint_angles}
\alias{compute_joint_angles}
\title{Compute Joint Angles from Rotation Frames}
\usage{
compute_joint_angles(rotation_frames)
}
\arguments{
\item{rotation_frames}{A 3-dimensional array of size 3x3xN, where each 3x3 slice represents a rotation matrix.}
}
\value{
A matrix of size N x 3, where each row contains the Euler angles (in degrees) corresponding to a 3x3 rotation matrix.
}
\description{
This function computes Euler angles from a series of 3x3 rotation matrices.
}
\examples{
# Create a 3x3x2 array of rotation matrices
rotation_frames <- array(c(
  1, 0, 0, 0, 1, 0, 0, 0, 1,  # Matrix 1
  0, -1, 0, 1, 0, 0, 0, 0, 1,  # Matrix 2
  0, 0, 1, 0, 0, 0, -1, 0, 0   # Matrix 3
), dim = c(3, 3, 3))
compute_joint_angles(rotation_frames)
}
